Move Add/remove liquidity to pools folder
