+++ src/pages/Pools/AddLiquidity/ConfirmAddModalBottom.tsx
@@ -2,11 +2,11 @@ import { Currency, CurrencyAmount, Fraction, Percent } from '@swapr/sdk'
 
 import { Text } from 'rebass'
 
-import { ButtonPrimary } from '../../components/Button'
-import { CurrencyLogo } from '../../components/CurrencyLogo'
-import { RowBetween, RowFixed } from '../../components/Row'
-import { Field } from '../../state/mint/actions'
-import { TYPE } from '../../theme'
+import { ButtonPrimary } from '../../../components/Button'
+import { CurrencyLogo } from '../../../components/CurrencyLogo'
+import { RowBetween, RowFixed } from '../../../components/Row'
+import { Field } from '../../../state/mint/actions'
+import { TYPE } from '../../../theme'
 
 export function ConfirmAddModalBottom({
   noLiquidity,
