+++ src/pages/Pools/AddLiquidity/ProtocolFeeBar.tsx
@@ -3,10 +3,10 @@ import { CurrencyAmount, Percent } from '@swapr/sdk'
 import { Text } from 'rebass'
 import { useTheme } from 'styled-components'
 
-import { AutoColumn } from '../../components/Column'
-import QuestionHelper from '../../components/QuestionHelper'
-import { AutoRow } from '../../components/Row'
-import { TYPE } from '../../theme'
+import { AutoColumn } from '../../../components/Column'
+import QuestionHelper from '../../../components/QuestionHelper'
+import { AutoRow } from '../../../components/Row'
+import { TYPE } from '../../../theme'
 
 export function ProtocolFeeBar({
   feePercentage,
