+++ src/pages/Pools/AddLiquidity/index.tsx
@@ -8,38 +8,40 @@ import { useParams } from 'react-router-dom'
 import { Text } from 'rebass'
 import { useTheme } from 'styled-components'
 
-import { ButtonPrimary } from '../../components/Button'
-import { BlueCard, OutlineCard } from '../../components/Card'
-import { AutoColumn, ColumnCenter } from '../../components/Column'
-import { CurrencyInputPanel } from '../../components/CurrencyInputPanel'
-import DoubleCurrencyLogo from '../../components/DoubleLogo'
-import { PageMetaData } from '../../components/PageMetaData'
-import { PoolLiquidityHeader } from '../../components/Pool/PoolLiquidityHeader'
-import { MinimalPositionCard } from '../../components/PositionCard'
-import Row, { RowBetween, RowFlat } from '../../components/Row'
-import TradePrice from '../../components/Swap/TradePrice'
-import TransactionConfirmationModal, { ConfirmationModalContent } from '../../components/TransactionConfirmationModal'
-import { PairState } from '../../data/Reserves'
-import { useActiveWeb3React } from '../../hooks'
-import { useCurrency } from '../../hooks/Tokens'
-import { ApprovalState, useApproveCallback } from '../../hooks/useApproveCallback'
-import { useWrappingToken } from '../../hooks/useContract'
-import { useNativeCurrency } from '../../hooks/useNativeCurrency'
-import { useRouter } from '../../hooks/useRouter'
-import useTransactionDeadline from '../../hooks/useTransactionDeadline'
-import { useWalletSwitcherPopoverToggle } from '../../state/application/hooks'
-import { Field } from '../../state/mint/actions'
-import { useDerivedMintInfo, useMintActionHandlers, useMintState } from '../../state/mint/hooks'
-import { useTransactionAdder } from '../../state/transactions/hooks'
-import { useIsExpertMode, useUserSlippageTolerance } from '../../state/user/hooks'
-import { TYPE } from '../../theme'
-import { calculateGasMargin, calculateSlippageAmount, getRouterContract } from '../../utils'
-import { currencyId } from '../../utils/currencyId'
-import { maxAmountSpend } from '../../utils/maxAmountSpend'
-import { calculateProtocolFee } from '../../utils/prices'
-import { wrappedCurrency } from '../../utils/wrappedCurrency'
-import AppBody from '../AppBody'
-import { Dots, Wrapper } from '../Pools/styleds'
+import { ButtonPrimary } from '../../../components/Button'
+import { BlueCard, OutlineCard } from '../../../components/Card'
+import { AutoColumn, ColumnCenter } from '../../../components/Column'
+import { CurrencyInputPanel } from '../../../components/CurrencyInputPanel'
+import DoubleCurrencyLogo from '../../../components/DoubleLogo'
+import { PageMetaData } from '../../../components/PageMetaData'
+import { PoolLiquidityHeader } from '../../../components/Pool/PoolLiquidityHeader'
+import { MinimalPositionCard } from '../../../components/PositionCard'
+import Row, { RowBetween, RowFlat } from '../../../components/Row'
+import TradePrice from '../../../components/Swap/TradePrice'
+import TransactionConfirmationModal, {
+  ConfirmationModalContent,
+} from '../../../components/TransactionConfirmationModal'
+import { PairState } from '../../../data/Reserves'
+import { useActiveWeb3React } from '../../../hooks'
+import { useCurrency } from '../../../hooks/Tokens'
+import { ApprovalState, useApproveCallback } from '../../../hooks/useApproveCallback'
+import { useWrappingToken } from '../../../hooks/useContract'
+import { useNativeCurrency } from '../../../hooks/useNativeCurrency'
+import { useRouter } from '../../../hooks/useRouter'
+import useTransactionDeadline from '../../../hooks/useTransactionDeadline'
+import { useWalletSwitcherPopoverToggle } from '../../../state/application/hooks'
+import { Field } from '../../../state/mint/actions'
+import { useDerivedMintInfo, useMintActionHandlers, useMintState } from '../../../state/mint/hooks'
+import { useTransactionAdder } from '../../../state/transactions/hooks'
+import { useIsExpertMode, useUserSlippageTolerance } from '../../../state/user/hooks'
+import { TYPE } from '../../../theme'
+import { calculateGasMargin, calculateSlippageAmount, getRouterContract } from '../../../utils'
+import { currencyId } from '../../../utils/currencyId'
+import { maxAmountSpend } from '../../../utils/maxAmountSpend'
+import { calculateProtocolFee } from '../../../utils/prices'
+import { wrappedCurrency } from '../../../utils/wrappedCurrency'
+import AppBody from '../../AppBody'
+import { Dots, Wrapper } from '../styleds'
 import { ConfirmAddModalBottom } from './ConfirmAddModalBottom'
 
 type CurrencySearchParams = {
