+++ src/pages/Pools/RemoveLiquidity/index.tsx
@@ -10,42 +10,44 @@ import { useParams } from 'react-router-dom'
 import { Box, Flex, Text } from 'rebass'
 import styled, { useTheme } from 'styled-components'
 
-import { ButtonConfirmed, ButtonError, ButtonPrimary } from '../../components/Button'
-import { LightCard } from '../../components/Card'
-import { AutoColumn, ColumnCenter } from '../../components/Column'
-import { CurrencyInputPanel } from '../../components/CurrencyInputPanel'
-import { CurrencyLogo } from '../../components/CurrencyLogo'
-import DoubleCurrencyLogo from '../../components/DoubleLogo'
-import { Option } from '../../components/Option'
-import { PageMetaData } from '../../components/PageMetaData'
-import { PoolLiquidityHeader } from '../../components/Pool/PoolLiquidityHeader'
-import { MinimalPositionCard } from '../../components/PositionCard'
-import Row, { RowBetween, RowFixed } from '../../components/Row'
-import Slider from '../../components/Slider'
-import { Dots } from '../../components/Swap/styleds'
-import TradePrice from '../../components/Swap/TradePrice'
-import TransactionConfirmationModal, { ConfirmationModalContent } from '../../components/TransactionConfirmationModal'
-import { useActiveWeb3React } from '../../hooks'
-import { useCurrency } from '../../hooks/Tokens'
-import { ApprovalState, useApproveCallback } from '../../hooks/useApproveCallback'
-import { usePairContract, useWrappingToken } from '../../hooks/useContract'
-import { useDebouncedChangeHandler } from '../../hooks/useDebouncedChangeHandler'
-import useIsArgentWallet from '../../hooks/useIsArgentWallet'
-import { useNativeCurrency } from '../../hooks/useNativeCurrency'
-import { useRouter } from '../../hooks/useRouter'
-import useTransactionDeadline from '../../hooks/useTransactionDeadline'
-import { useWalletSwitcherPopoverToggle } from '../../state/application/hooks'
-import { Field } from '../../state/burn/actions'
-import { useBurnActionHandlers, useBurnState, useDerivedBurnInfo } from '../../state/burn/hooks'
-import { useTransactionAdder } from '../../state/transactions/hooks'
-import { useUserSlippageTolerance } from '../../state/user/hooks'
-import { StyledInternalLink, TYPE } from '../../theme'
-import { calculateGasMargin, calculateSlippageAmount, getRouterContract } from '../../utils'
-import { currencyId } from '../../utils/currencyId'
-import { calculateProtocolFee } from '../../utils/prices'
-import { wrappedCurrency } from '../../utils/wrappedCurrency'
-import AppBody from '../AppBody'
-import { Wrapper } from '../Pools/styleds'
+import { ButtonConfirmed, ButtonError, ButtonPrimary } from '../../../components/Button'
+import { LightCard } from '../../../components/Card'
+import { AutoColumn, ColumnCenter } from '../../../components/Column'
+import { CurrencyInputPanel } from '../../../components/CurrencyInputPanel'
+import { CurrencyLogo } from '../../../components/CurrencyLogo'
+import DoubleCurrencyLogo from '../../../components/DoubleLogo'
+import { Option } from '../../../components/Option'
+import { PageMetaData } from '../../../components/PageMetaData'
+import { PoolLiquidityHeader } from '../../../components/Pool/PoolLiquidityHeader'
+import { MinimalPositionCard } from '../../../components/PositionCard'
+import Row, { RowBetween, RowFixed } from '../../../components/Row'
+import Slider from '../../../components/Slider'
+import { Dots } from '../../../components/Swap/styleds'
+import TradePrice from '../../../components/Swap/TradePrice'
+import TransactionConfirmationModal, {
+  ConfirmationModalContent,
+} from '../../../components/TransactionConfirmationModal'
+import { useActiveWeb3React } from '../../../hooks'
+import { useCurrency } from '../../../hooks/Tokens'
+import { ApprovalState, useApproveCallback } from '../../../hooks/useApproveCallback'
+import { usePairContract, useWrappingToken } from '../../../hooks/useContract'
+import { useDebouncedChangeHandler } from '../../../hooks/useDebouncedChangeHandler'
+import useIsArgentWallet from '../../../hooks/useIsArgentWallet'
+import { useNativeCurrency } from '../../../hooks/useNativeCurrency'
+import { useRouter } from '../../../hooks/useRouter'
+import useTransactionDeadline from '../../../hooks/useTransactionDeadline'
+import { useWalletSwitcherPopoverToggle } from '../../../state/application/hooks'
+import { Field } from '../../../state/burn/actions'
+import { useBurnActionHandlers, useBurnState, useDerivedBurnInfo } from '../../../state/burn/hooks'
+import { useTransactionAdder } from '../../../state/transactions/hooks'
+import { useUserSlippageTolerance } from '../../../state/user/hooks'
+import { StyledInternalLink, TYPE } from '../../../theme'
+import { calculateGasMargin, calculateSlippageAmount, getRouterContract } from '../../../utils'
+import { currencyId } from '../../../utils/currencyId'
+import { calculateProtocolFee } from '../../../utils/prices'
+import { wrappedCurrency } from '../../../utils/wrappedCurrency'
+import AppBody from '../../AppBody'
+import { Wrapper } from '../styleds'
 
 const StyledInternalLinkText = styled(TYPE.Body)`
   display: flex;
